# deploy_v2/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GroupBuy Backend API
  description: 'REST API for the GroupBuy platform — group purchasing management.


    **Authentication**: Bearer JWT token. Obtain via `POST /auth/login`.


    **Roles**: regular users can manage their own data; admin users can manage all
    resources.


    Interactive docs: `/docs` (Swagger UI) · `/redoc` (ReDoc)'
  contact:
    name: GroupBuy Team
  version: 2.0.0
paths:
...
  /auth/login:
    post:
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /users/{user_id}:
    get:
      summary: Get User
      operationId: get_user_users__user_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update User
      operationId: update_user_users__user_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete User
      operationId: delete_user_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /categories:
    get:
      summary: List Categories
      operationId: list_categories_categories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryOut'
                type: array
                title: Response List Categories Categories Get
    post:
      summary: Create Category
      operationId: create_category_categories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /categories/{cat_id}:
    delete:
      summary: Delete Category
      operationId: delete_category_categories__cat_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: cat_id
        in: path
        required: true
        schema:
          type: integer
          title: Cat Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /procurements:
    get:
      summary: List Procurements
      operationId: list_procurements_procurements_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: city
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: City
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcurementOut'
                title: Response List Procurements Procurements Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Procurement
      operationId: create_procurement_procurements_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcurementCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcurementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /procurements/{proc_id}:
    get:
      summary: Get Procurement
      operationId: get_procurement_procurements__proc_id__get
      parameters:
      - name: proc_id
        in: path
        required: true
        schema:
          type: integer
          title: Proc Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcurementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Procurement
      operationId: update_procurement_procurements__proc_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: proc_id
        in: path
        required: true
        schema:
          type: integer
          title: Proc Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcurementUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcurementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Procurement
      operationId: delete_procurement_procurements__proc_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: proc_id
        in: path
        required: true
        schema:
          type: integer
          title: Proc Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /payments:
    get:
      summary: List Payments
      operationId: list_payments_payments_get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentOut'
                title: Response List Payments Payments Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Payment
      operationId: create_payment_payments_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /admin/stats:
    get:
      summary: Admin Stats
      operationId: admin_stats_admin_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
components:
  schemas:
    CategoryCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
        parent_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Id
        icon:
          type: string
          title: Icon
          default: ''
      type: object
      required:
      - name
      title: CategoryCreate
    CategoryOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        parent_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Id
        icon:
          type: string
          title: Icon
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - description
      - parent_id
      - icon
      - is_active
      - created_at
      title: CategoryOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    PaymentCreate:
      properties:
        payment_type:
          type: string
          title: Payment Type
        amount:
          type: number
          title: Amount
        procurement_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Procurement Id
        description:
          type: string
          title: Description
          default: ''
      type: object
      required:
      - payment_type
      - amount
      title: PaymentCreate
    PaymentOut:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        procurement_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Procurement Id
        payment_type:
          type: string
          title: Payment Type
        amount:
          type: number
          title: Amount
        status:
          type: string
          title: Status
        description:
          type: string
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - user_id
      - procurement_id
      - payment_type
      - amount
      - status
      - description
      - created_at
      title: PaymentOut
    ProcurementCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
          default: ''
        category_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
        city:
          type: string
          title: City
          default: ''
        delivery_address:
          type: string
          title: Delivery Address
          default: ''
        target_amount:
          type: number
          title: Target Amount
        stop_at_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Stop At Amount
        unit:
          type: string
          title: Unit
          default: units
        price_per_unit:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Per Unit
        commission_percent:
          type: number
          title: Commission Percent
          default: 0.0
        deadline:
          type: string
          format: date-time
          title: Deadline
        image_url:
          type: string
          title: Image Url
          default: ''
        is_featured:
          type: boolean
          title: Is Featured
          default: false
      type: object
      required:
      - title
      - target_amount
      - deadline
      title: ProcurementCreate
    ProcurementOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
        organizer_id:
          type: integer
          title: Organizer Id
        organizer_username:
          type: string
          title: Organizer Username
        city:
          type: string
          title: City
        delivery_address:
          type: string
          title: Delivery Address
        target_amount:
          type: number
          title: Target Amount
        current_amount:
          type: number
          title: Current Amount
        stop_at_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Stop At Amount
        unit:
          type: string
          title: Unit
        price_per_unit:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Per Unit
        commission_percent:
          type: number
          title: Commission Percent
        status:
          type: string
          title: Status
        deadline:
          type: string
          format: date-time
          title: Deadline
        image_url:
          type: string
          title: Image Url
        is_featured:
          type: boolean
          title: Is Featured
        participant_count:
          type: integer
          title: Participant Count
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - title
      - description
      - category_id
      - organizer_id
      - organizer_username
      - city
      - delivery_address
      - target_amount
      - current_amount
      - stop_at_amount
      - unit
      - price_per_unit
      - commission_percent
      - status
      - deadline
      - image_url
      - is_featured
      - participant_count
      - created_at
      - updated_at
      title: ProcurementOut
    ProcurementUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        delivery_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Delivery Address
        target_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Target Amount
        stop_at_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Stop At Amount
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        price_per_unit:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Per Unit
        commission_percent:
          anyOf:
          - type: number
          - type: 'null'
          title: Commission Percent
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_featured:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Featured
      type: object
      title: ProcurementUpdate
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      title: TokenResponse
...
    UserOut:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        is_active:
          type: boolean
          title: Is Active
        is_admin:
          type: boolean
          title: Is Admin
        balance:
          type: number
          title: Balance
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - username
      - email
      - is_active
      - is_admin
      - balance
      - created_at
      title: UserOut
    UserUpdate:
      properties:
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer